refactor: functionallity into gitfloat
