+++ source-code/website/src/pages/editor/@host/@owner/@repository/Layout.tsx
@@ -44,7 +44,6 @@ export function Layout(props: { children: JSXElement }) {
 	return (
 		<RootLayout>
 			<div class="pt-4 pb-16 w-full flex flex-col grow">
-				<SignInBanner />
 				<div class="flex items-center space-x-4 pt-5">
 					<Breadcrumbs />
 					<BranchMenu />
@@ -64,7 +63,6 @@ export function Layout(props: { children: JSXElement }) {
 					</div>
 					<div class="flex gap-2">
 						<SearchInput placeholder="Search ..." handleChange={handleSearchText} />
-						<HasChangesAction />
 					</div>
 				</div>
 				{/* <hr class="h-px w-full bg-outline-variant my-2"> </hr> */}
@@ -151,89 +149,6 @@ function BranchMenu() {
 					</For> */}
 			</sl-menu>
 		</sl-dropdown>
-	)
-}
-
-/** Actions that can be conducted if a commit has been made but not pushed yet. */
-function HasChangesAction() {
-	const { unpushedChanges, setFsChange, setLastPush, routeParams, fs, setLastPullTime } =
-		useEditorState()
-
-	const [latestChange, setLatestChange] = createSignal<Date>()
-	const [showPulse, setShowPulse] = createSignal(false)
-	const [isLoading, setIsLoading] = createSignal(false)
-	const [localStorage] = useLocalStorage()
-
-	createEffect(() => {
-		if (unpushedChanges()) {
-			setLatestChange(new Date())
-		}
-	})
-
-	// show the pulse if less than X seconds ago a change has been conducted
-	const interval = setInterval(() => {
-		const _latestChange = latestChange()
-		if (_latestChange === undefined) {
-			return setShowPulse(false)
-		}
-		const eightSecondsAgo = subSeconds(new Date(), 8)
-		return setShowPulse(isAfter(_latestChange, eightSecondsAgo))
-	}, 1000)
-
-	onCleanup(() => clearInterval(interval))
-
-	async function triggerPushChanges() {
-		if (localStorage?.user === undefined) {
-			return showToast({
-				title: "Failed to push changes",
-				message: "Please login first",
-				variant: "warning",
-			})
-		}
-		setIsLoading(true)
-		const [, exception] = await pushChanges({
-			fs: fs(),
-			routeParams: routeParams(),
-			user: localStorage.user,
-			setFsChange,
-			setLastPush,
-			setLastPullTime,
-		})
-		setIsLoading(false)
-		telemetryBrowser.capture("push changes", {
-			owner: routeParams().owner,
-			repository: routeParams().repository,
-			sucess: exception === undefined,
-		})
-		if (exception) {
-			return showToast({
-				title: "Failed to push changes",
-				message: "Please try again or file a bug. " + exception,
-				variant: "danger",
-			})
-		} else {
-			setHasPushedChanges(true)
-			return showToast({
-				title: "Changes have been pushed",
-				variant: "success",
-			})
-		}
-	}
-
-	return (
-		<sl-button
-			prop:disabled={(unpushedChanges() ?? []).length === 0}
-			onClick={triggerPushChanges}
-			prop:loading={isLoading()}
-			prop:size={"small"}
-		>
-			Push changes
-			<Show when={(unpushedChanges() ?? []).length > 0}>
-				<sl-badge prop:pill={true} prop:pulse={showPulse() ? true : false}>
-					{(unpushedChanges() ?? []).length}
-				</sl-badge>
-			</Show>
-		</sl-button>
 	)
 }
 
@@ -422,188 +337,5 @@ function StatusFilter() {
 				</For>
 			</div>
 		</sl-select>
-	)
-}
-
-function SignInBanner() {
-	const { userIsCollaborator, githubRepositoryInformation, currentBranch, routeParams } =
-		useEditorState()
-	const [localStorage] = useLocalStorage()
-	const [isLoading, setIsLoading] = createSignal(false)
-
-	let alert: SlAlert | undefined
-
-	createEffect(() => {
-		// workaround for shoelace animation
-		if (userIsCollaborator() === false) {
-			setTimeout(() => {
-				alert?.show()
-			}, 50)
-		} else {
-			alert?.hide()
-		}
-	})
-
-	let signInDialog: SlDialog | undefined
-
-	function onSignIn() {
-		signInDialog?.show()
-	}
-
-	async function handleFork() {
-		setIsLoading(true)
-		if (localStorage.user === undefined) {
-			return
-		}
-		const response = await github.rest.repos.createFork({
-			owner: routeParams().owner,
-			repo: routeParams().repository,
-		})
-		telemetryBrowser.capture("fork created", {
-			owner: routeParams().owner,
-			repository: routeParams().repository,
-			sucess: response.status === 202,
-		})
-		if (response.status === 202) {
-			showToast({
-				variant: "success",
-				title: "The Fork has been created.",
-				message: `Don't forget to open a pull request`,
-			})
-			setIsLoading(false)
-			// full name is owner/repo
-			return navigate(`/editor/github.com/${response.data.full_name}`)
-		} else {
-			showToast({
-				variant: "danger",
-				title: "The creation of the fork failed.",
-				message: `Please try it again or report a bug`,
-			})
-			return response
-		}
-	}
-	return (
-		<>
-			<Switch>
-				<Match when={localStorage?.user === undefined}>
-					<Banner
-						variant="info"
-						message="The repository has been cloned locally. You must sign in with your GitHub account to commit and push changes."
-					>
-						<sl-button onClick={onSignIn} prop:variant="primary">
-							Sign in with GitHub
-							<CibGithub slot="suffix" />
-						</sl-button>
-					</Banner>
-				</Match>
-				<Match
-					when={
-						userIsCollaborator.error === undefined &&
-						userIsCollaborator.loading === false &&
-						userIsCollaborator() === false &&
-						localStorage?.user
-					}
-				>
-					<Banner
-						variant="info"
-						message={`
-            You do not have write access to ${routeParams().owner}/${
-							routeParams().repository
-						}. Fork this project to make changes.`}
-					>
-						<sl-button onClick={handleFork} prop:variant="primary" prop:loading={isLoading()}>
-							<div slot="prefix">
-								<svg width="1.2em" height="1.2em" viewBox="0 0 16 16">
-									<path
-										fill="currentColor"
-										d="M5 5.372v.878c0 .414.336.75.75.75h4.5a.75.75 0 0 0 .75-.75v-.878a2.25 2.25 0 1 1 1.5 0v.878a2.25 2.25 0 0 1-2.25 2.25h-1.5v2.128a2.251 2.251 0 1 1-1.5 0V8.5h-1.5A2.25 2.25 0 0 1 3.5 6.25v-.878a2.25 2.25 0 1 1 1.5 0ZM5 3.25a.75.75 0 1 0-1.5 0a.75.75 0 0 0 1.5 0Zm6.75.75a.75.75 0 1 0 0-1.5a.75.75 0 0 0 0 1.5Zm-3 8.75a.75.75 0 1 0-1.5 0a.75.75 0 0 0 1.5 0Z"
-									/>
-								</svg>
-							</div>
-							Fork this project
-						</sl-button>
-					</Banner>
-				</Match>
-				<Match
-					when={
-						githubRepositoryInformation.error === undefined &&
-						githubRepositoryInformation.loading === false &&
-						hasPushedChanges() &&
-						githubRepositoryInformation()?.data.fork
-					}
-				>
-					<Banner
-						variant="success"
-						message={`You are working in a forked project. Please make a "pull request" to transfer your changes to the parent project:
-							"${githubRepositoryInformation()?.data.parent?.full_name}"`}
-					>
-						<sl-button
-							prop:target="_blank"
-							prop:href={`https://github.com/${
-								githubRepositoryInformation()?.data.parent?.full_name
-							}/compare/${currentBranch()}...${githubRepositoryInformation()?.data.owner.login}:${
-								githubRepositoryInformation()?.data.name
-							}:${currentBranch()}?expand=1;title=Update%20translations;body=Describe%20the%20changes%20you%20have%20conducted%20here%0A%0APreview%20the%20messages%20on%20https%3A%2F%2Finlang.com%2Fgithub.com%2F${
-								(currentPageContext.routeParams as EditorRouteParams).owner
-							}%2F${(currentPageContext.routeParams as EditorRouteParams).repository}%20.`}
-							prop:variant="success"
-							// ugly workaround to close  the banner
-							// after the button has been clicked
-							onClick={() => {
-								telemetryBrowser.capture("open pull request", {
-									owner: routeParams().owner,
-									repository: routeParams().repository,
-								})
-								setHasPushedChanges(false)
-							}}
-						>
-							<div slot="prefix">
-								<svg width="1em" height="1em" viewBox="0 0 24 24">
-									<path
-										fill="currentColor"
-										d="M16 19.25a3.25 3.25 0 1 1 6.5 0a3.25 3.25 0 0 1-6.5 0Zm-14.5 0a3.25 3.25 0 1 1 6.5 0a3.25 3.25 0 0 1-6.5 0Zm0-14.5a3.25 3.25 0 1 1 6.5 0a3.25 3.25 0 0 1-6.5 0ZM4.75 3a1.75 1.75 0 1 0 .001 3.501A1.75 1.75 0 0 0 4.75 3Zm0 14.5a1.75 1.75 0 1 0 .001 3.501A1.75 1.75 0 0 0 4.75 17.5Zm14.5 0a1.75 1.75 0 1 0 .001 3.501a1.75 1.75 0 0 0-.001-3.501Z"
-									/>
-									<path
-										fill="currentColor"
-										d="M13.405 1.72a.75.75 0 0 1 0 1.06L12.185 4h4.065A3.75 3.75 0 0 1 20 7.75v8.75a.75.75 0 0 1-1.5 0V7.75a2.25 2.25 0 0 0-2.25-2.25h-4.064l1.22 1.22a.75.75 0 0 1-1.061 1.06l-2.5-2.5a.75.75 0 0 1 0-1.06l2.5-2.5a.75.75 0 0 1 1.06 0ZM4.75 7.25A.75.75 0 0 1 5.5 8v8A.75.75 0 0 1 4 16V8a.75.75 0 0 1 .75-.75Z"
-									/>
-								</svg>
-							</div>
-							Create pull request
-						</sl-button>
-					</Banner>
-				</Match>
-			</Switch>
-			{/* <sl-button onClick={handlesncForking}>can i fork this thing</sl-button> */}
-			<SignInDialog
-				githubAppClientId={publicEnv.PUBLIC_GITHUB_APP_CLIENT_ID}
-				ref={signInDialog!}
-				onClickOnSignInButton={() => {
-					// hide the sign in dialog to increase UX when switching back to this window
-					signInDialog?.hide()
-				}}
-			/>
-		</>
-	)
-}
-
-function Banner(props: {
-	variant: SemanticColorTokens[number]
-	message: string
-	children: JSXElement
-}) {
-	let alert: SlAlert | undefined
-	return (
-		<sl-alert
-			prop:variant={props.variant === "info" ? "primary" : props.variant}
-			ref={alert}
-			prop:open={true}
-		>
-			<Icon name={props.variant} slot="icon" />
-			<div class="flex space-x-4 items-center">
-				<p class="grow">{props.message}</p>
-				{props.children}
-			</div>
-		</sl-alert>
 	)
 }
